added map search bar and skills to a new component
